Allow the user to select SSL modes.
pq.SetSslMode now allows the user to alter the default setting of always
attempting SSL and silently accepting the Postgres server's wishes. This
can now be set to SslModeNever to never use SSL or SslModeForced to fail
if the SSLRequest is denied. Set to SslModeAttempt by default.

pq.SetSslVerifyPeer can be set to true or false to control whether or
not the TLS connection should verify that the peer's certificate is
signed by a CA in the keychain and that it was actually issued for that
host. Set to 'true' by default.